Added parsing to intermediate language
- Reorganized some files, fixed some import bugs.

Added a new intermediate language called a ParsedGrammarTree. A
ParsedGrammarTree will be used to get the actual AST generated by the
language. For now a ParsedGrammarTree is language-agnostic, only
determined by the formal grammar. It will be up to the AST to ensure
that the ParsedGrammarTree is in the correct format.

At least that's the current plan...